refactor: move ContactForm out of Home into file with 'use client'
