+++ apps/web/src/app/page.tsx
@@ -1,20 +1,14 @@
-"use client";
-import ContactButton from "@/components/common/contact/ContactButton";
-import ContactForm from "@/components/common/contact/ContactForm";
-import Hero from "@/components/common/Hero";
-import { useState } from "react";
+import Hero from "@/components/sections/Hero";
+import ContactSection from "@/components/sections/ContactSection";
 
 export default function Home() {
-  const [isContactOpen, setIsContactOpen] = useState(false);
-
   return (
     <>
-      <Hero />
-      <ContactButton onClick={() => setIsContactOpen(true)} />
-      <ContactForm
-        isOpen={isContactOpen}
-        onClose={() => setIsContactOpen(false)}
-      />
+      <div>
+        <Hero />
+      </div>
+
+      <ContactSection />
     </>
   );
 }
